Moved classes in external windows to their own files, small fixes, added image selector text box
