+++ Phindr3D-Python/src/GUI/windows/__init__.py
@@ -1,0 +1,12 @@
+import sys, os.path
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
+from .colorchannelWindow import *
+from .extractWindow import *
+from .featurefilegroupingwindow import *
+from .helperclasses import *
+from .interactive_click import *
+from .paramWindow import *
+from .regexWindow import *
+from .resultsWindow import *
+from .segmentationWindow import *
